Move ArtistTreatmentCompleted and ArtistTreatmentIndex imports to artist folder; fix totalTime calculation in ArtistTreatmentPickMoreRenderModal
